+++ src/cllazyfile/p21HeaderSectionReader.cc
@@ -36,7 +36,7 @@ const namedLazyInstance p21HeaderSectionReader::nextInstance()
     namedLazyInstance i;
     static instanceID nextFreeInstance = 4; // 1-3 are reserved per 10303-21
 
-    i.loc.begin = _file.tellg();
+    i.loc.begin = static_cast<long>(_file.tellg());
     i.loc.section = _sectionID;
     skipWS();
     if(i.loc.begin <= 0) {
